NMEAProtocol

Documentation for NMEAProtocol.

NMEAProtocol

Public Methods

Private Methods

Types/Structs

Public Methods

NMEAProtocol.parse — Function

NMEAProtocol.parse(::NMEAPacket, nmeastring)
NMEAProtocol.parse(::NMEAPacket{AbstractNMEAMessage}, nmeastring)

parses a string into a NMEAPacket. specifying the sub-type of nmea string improves performance. But the type will be looked up if it is not specified.

NMEA Type Structs

NMEAProtocol.GGA — Type

GGA(msg)

NMEA-0183 message: GGA

Time, position, and fix related data.

Format

    $<TalkerID>GGA,<Timestamp>,<Lat>,<N/S>,<Long>,<E/W>,<GPSQual>,<Sats>,<HDOP>,<Alt>,<AltVal>,<GeoSep>,<GeoVal>,<DGPSAge>,<DGPSRef>*<checksum><CR><LF>

Example

    $GPGGA,172814.0,3723.46587704,N,12202.26957864,W,2,6,1.2,18.893,M,-25.669,M,2.0 0031*4F

NMEAProtocol.GSA — Type

GSA

NMEAProtocol.DTM — Type

DTM

NMEAProtocol.GBS — Type

GBS

NMEAProtocol.GLL — Type

GLL

NMEAProtocol.GSV — Type

GSV

NMEAProtocol.GST — Type

GST

NMEAProtocol.RMC — Type

RMC

NMEAProtocol.VTG — Type

VTG

NMEAProtocol.ZDA — Type

ZDA

Private Methods

Types/Structs

NMEAProtocol.SYSTEM — Module

SYSTEM

NMEA sentences include a Talker Identifier (ID) after the “$” sign. The talker ID is a two-character prefix that identifies the type of the transmitting unit.

Standard Systems are:

GA European Global Navigation System (Galileo

GB BeiDou Navigation Satellite System (BDS)

GI Navigation Indian Constellatiozn (NavIC)

GL Globalnaya Navigazionnaya Sputnikovaya Sistema (GLONASS)

GN Global Navigation Satellite System (GNSS)

GP Global Positioning System (GPS)

GQ Quasi-Zenith Satellite System (QZSS)

There are aditional IDs for non-standard sources.

NMEAProtocol.NMEAPacket — Type

NMEAPacket
NMEAPacket{<:AbstractNMEAMessage}

wrapping type of all nmea messages. contains information from the string header. The satelite system used, the time the string was parsed and if the checksum matches the message hash. and of course the NMEA string for the data.

Parsing

NMEAProtocol._to_system — Function

_to_system(id)

lookup the system being used based on the 2 character talker id

NMEAProtocol._to_type — Function

_to_type(format)

lookup the type of nmea string based on the 3 character message format identifier

NMEAProtocol._to_int — Function

_to_int(item)

Converts a string representing an integer to a Int, if parse fails it defaults to 0.

NMEAProtocol._to_float — Function

tofloat(item) Converts a string representing an float to a Float64, if parse fails it defaults to 0.0.

NMEAProtocol._to_decimal_deg — Function

todecimal_deg(dms, hemi)

Converts a string representing longitude or latitude in degrees and minutes (DDDMM.mmm [N|S|E|W]) to decimal degrees (±DD.ddd). if either arguments are nothing returns 0.0

NMEAProtocol._to_time — Function

totime(timestamp)

Converts a string in the format HHMMSS.sss to a Dates.Time object. If the argument is nothing the current time in UTC will be returned.

NMEAProtocol._to_distance — Function

todistance(dist, unit)

convert a distance in a given unit to a distance in meters

NMEAProtocol._to_speed — Function

tospeed(velocity, unit)

convert a speed in a given unit to a velocity in meters/s

NMEAProtocol._hash_msg — Function

_hash_msg(message)

perform an xor hash of a string.

NMEAProtocol.SYSTEM

NMEAProtocol.DTM

NMEAProtocol.GBS

NMEAProtocol.GGA

NMEAProtocol.GLL

NMEAProtocol.GSA

NMEAProtocol.GST

NMEAProtocol.GSV

NMEAProtocol.NMEAPacket

NMEAProtocol.RMC

NMEAProtocol.VTG

NMEAProtocol.ZDA

NMEAProtocol._hash_msg

NMEAProtocol._to_decimal_deg

NMEAProtocol._to_distance

NMEAProtocol._to_float

NMEAProtocol._to_int

NMEAProtocol._to_speed

NMEAProtocol._to_system

NMEAProtocol._to_time

NMEAProtocol._to_type

NMEAProtocol.parse